Problem: "waiting" interface was not unified.
The various thing that could put a task to
sleep didn't use a common interface to do so.

Solution: Implement WaitHandler.
Task activity is now tracker through those
WaitHandler. When a task yield to an other
coroutine, a WaitHandler is attached to the
parent task. When the child coroutine finishes,
the handler is detached, and the parent
coroutine will be marked active.
Same for ell::sleep() or wait on ConditionVariable.

Note: Code is not clean.